added routes folder and cleaned up backend code
