+++ routes/Login.js
@@ -1,0 +1,26 @@
+const express = require('express')
+const User = require('../models/User')
+const bcrypt = require('bcryptjs')
+const jwt = require('jsonwebtoken')
+const router = express.Router()
+
+router.post('/', async (req, res) => {
+    const user = await User.findOne({
+        email: req.body.email
+    })
+
+    const isPasswordValid = await bcrypt.compare(req.body.password, user.password)
+
+    if (isPasswordValid) {
+        const token = jwt.sign({
+            name: user.name,
+            email: user.email
+        }, 'secret1427')
+
+        return res.json({ status: 'ok', user: token })
+    } else {
+        return res.json({ status: 'error', user: false })
+    }
+})
+
+module.exports = router
